update clock bpm to float, disable watchdog causing unable to upload

+++ include/clock.h
@@ -14,7 +14,7 @@ struct Clock
 {
     float interval;
     float correction;
-    uint32_t bpm;
+    float bpm;
 
     TickCallback onTick;
 };
